v3.0.0
Add checkHeliPos and checkInput fun in app.cpp

+++ src/shoot.cpp
@@ -6,7 +6,7 @@ void shoot::shootStatus(int X, std::string board[][column], int Right)
 
 {
 
-    for (size_t i = 1; i < Right; i++)
+    for (int i = 1; i < Right; i++)
     {
         board[X][i] = "   ";
     }
